second commit

+++ src/main/java/org/example/model/game/Car.java
@@ -1,7 +1,7 @@
 package org.example.model.game;
 
 public class Car extends Rectangle {
-    public Car(float x1, float x2, float y1, float y2) {
+    public Car(int x1, int x2, int y1, int y2) {
         super(x1, x2, y1, y2);
     }
 }
