+++ src/main/java/org/example/model/game/Log.java
@@ -1,7 +1,7 @@
 package org.example.model.game;
 
 public class Log extends Rectangle {
-    public Log(float x1, float x2, float y1, float y2) {
+    public Log(int x1, int x2, int y1, int y2) {
         super(x1, x2, y1, y2);
     }
 }
